When duplicate Main_Loop automatic GO, delete the MainLoop component && Add feedback when Systems list updated during runtime.

+++ Defenses_immunitaires/Assets/Scripts/MainLoop/MainLoopInspector.cs
@@ -4,6 +4,7 @@ using UnityEditor;
 [CustomEditor(typeof(MainLoop))]
 public class MainLoopInspector : Editor {
 	private SerializedProperty _systemFiles;
+	private SerializedProperty _systemFilesChanged;
 	private GUILayoutOption _minButtonWidth;
 	private GUIContent _addButton;
 	private GUIContent _moveUpButton;
@@ -11,7 +12,13 @@ public class MainLoopInspector : Editor {
 	private GUIContent _removeButton;
 
 	private void OnEnable(){
+		if (Object.FindObjectsOfType<MainLoop>().Length > 1) { // due to duplicate hotkeys & option on GO
+			DestroyImmediate(target);
+			return;
+		}
+
 		_systemFiles = serializedObject.FindProperty("_systemFiles");
+		_systemFilesChanged = serializedObject.FindProperty("_systemFilesChanged");
 		_minButtonWidth = GUILayout.MinWidth(20f);
 		_addButton = new GUIContent("+", "add");
 		_moveUpButton = new GUIContent("\u25B2", "move up");
@@ -27,27 +34,33 @@ public class MainLoopInspector : Editor {
 			//---------------------------------------------------------------------------------------------------------------
 			SerializedProperty sp = _systemFiles.GetArrayElementAtIndex(i);
 			MonoScript script = (MonoScript) EditorGUILayout.ObjectField(sp.objectReferenceValue, typeof(MonoScript), false);
-			if (script != null && script.GetClass().IsSubclassOf(typeof(UECS.System)))
+			if (sp.objectReferenceValue != script && script != null && script.GetClass().IsSubclassOf(typeof(UECS.System))) {
 				sp.objectReferenceValue = script;
+				_systemFilesChanged.boolValue = true;
+			}
 			//---------------------------------------------------------------------------------------------------------------
 			if (GUILayout.Button (_moveUpButton, _minButtonWidth)) {
-				_systemFiles.MoveArrayElement (i, i - 1);
+				_systemFiles.MoveArrayElement(i, i - 1);
+				_systemFilesChanged.boolValue = true;
 			}
 			if (GUILayout.Button (_moveDownButton, _minButtonWidth)) {
 				_systemFiles.MoveArrayElement (i, i + 1);
+				_systemFilesChanged.boolValue = true;
 			}
-			if (GUILayout.Button (_removeButton, _minButtonWidth)) {
+			if (!Application.isPlaying && GUILayout.Button (_removeButton, _minButtonWidth)) {
 				_systemFiles.GetArrayElementAtIndex(i).objectReferenceValue = null;
 				_systemFiles.DeleteArrayElementAtIndex(i);
+				_systemFilesChanged.boolValue = true;
 			}
 			//---------------------------------------------------------------------------------------------------------------
 			EditorGUILayout.EndHorizontal();
 		}
 
-		if (GUILayout.Button (_addButton, EditorStyles.miniButtonLeft)) {
+		if (!Application.isPlaying && GUILayout.Button (_addButton, EditorStyles.miniButtonLeft)) {
 			int index = _systemFiles.arraySize;
 			_systemFiles.InsertArrayElementAtIndex(index);
 			_systemFiles.GetArrayElementAtIndex(index).objectReferenceValue = null;
+			_systemFilesChanged.boolValue = true;
 		}
 
 		EditorGUILayout.HelpBox ("Set all the systems you need", MessageType.Info, true);
